FLUID-2136: When we remove a row we also remove it's progressor.
Note: we're still leaving behind the progressor when a file has uploaded successfully even though it will never be used again. The problem with removing the progressor when a file is successful is that the progressor will be removed before it gets a chance to display 100% progress. There is some rethinking of the flow of the Uploader that could fix this but it doesn't seem worth messing with. 

svn path=/fluid/components/trunk/; revision=6329