FLUID-2297: removed old code from Progress, reran tests
In Progress changed this:

	that.refresh = function (newProgressBarElm) {
    	if (newProgressBarElm) {
        	that.progressBar = newProgressBarElm;
    	}
    
    	repositionIndicator(that);
	};

To this: 

	that.refresh = function () {
    	repositionIndicator(that);
	};

And removed the code that reset the width when repositioning an element, as you would want to reset the indicators width explicitly not when repositioning the element:

	var repositionIndicator = function (that) {
        that.indicator.css("top", that.progressBar.position().top)
            .css("left", 0)
            .height(that.progressBar.height())
            .width(that.options.minWidth);
    };
    
Also removed the element argument from the one place in Uplaoder.js that calls Progress.refresh(). It would never pass a different argument than the one that it was initailized with.

svn path=/fluid/components/trunk/; revision=6607