FLUID-1811: Added error messages for when a file queue fails.
Please note that there is some strange behavior in SWFUpload:
if the limit on the number of files that can be uploaded is 3 and the user attempts to select 4 files, instead of queing the first 3 files and then throwing an error, SWFUpload just throws an error and no files are queued. Fun! 

svn path=/fluid/infusion/trunk/; revision=7184