FLUID-3886: Convert all file types to HTML5 mime types
In version 1.3, the HTML5 version of the Uploader couldn't exclude certain file
types.  The file type option in the queue setting is now converted into valid
HTML5 mime types when the HTML5 uploader is in use.  A simple fix to repair a
broken unit test in the file queue has also been added.  See FLUID-4053.