FLUID-4025: Adding to FSS flex container styles
Added a complete set of percentage based flex container styles. This is based on the "flexible-columns.css" from uPortal sent by Gary to the fluid-work list ( http://old.nabble.com/FSS-expanded-flexible-columns-to31098851.html ). I modified it to conform with FSS flex container declarations. Note that I did not incorporate the IE 7 hack for 50% because it appears to be due to an issue with the browser's rounding at various window sizes. I was able to reproduce the issue both with and without the hack.