FLUID-4163: Added more integration tests
Added integration tests for the uploader instantiated as a subcomponent, as well as, a subcomponent demanding additional external options.

To summarize, we now have integration tests for the single-file, SWF, and HTML5 uploading strategies.  We have covered tests for plain instantiation, instantiation as a subcomponent, and instantion as a subcomponent with external demands.