FLUID-4150: Beginning refactoring js of test page
Pulling the js out of the html file and into it's own js file. This is the first step into getting it ready to be used as a demo for linearization. Still need to adjust the design of the option controls a bit, and make a component to handle the switching. Finally I'll promote the example out of the manual tests and into the fss demos.