FLUID-4206: Removed unneed jquery.ui.core.css file
The jquery.ui.core.css file was in the demos/lib directory. It is unneeded because the demo-theme handles the necessary styling. Rather than update it, I just removed it. I also fixed up the imports in the linearize.html file that had been calling it. And moved the import of the demo theme to the correct spot in the file.