FLUID-4209: First try at getting ToC working again
Made a bunch of changes and now have the ToC working again without RSF:ID's and in more of a modern style. Lots of cleanup still to do and there's still some refinement to do with generating all the headings generated correctly. The code could also use some optimization, as it's a bit slow at the moment. Of course, all the unit tests are broken, they will need to be updated and expanded on.