FLUID-4209: Step 1 of ToC refactoring from review
Antranig was nice enough to code review my work on the tableOfContents component(s). Step 1 of the refactoring, many more to come.

I've switched from fluid.each to fluid.transform in the fluid.tableOfContents.finalInit. Also changed the variables from tocAnchor, anchors, and etc. to more specifically illustrate that it is just an array of the urls.