FLUID-4209: Step 2 of ToC refactoring from review
Pulled out the code that had been in the fluid.transform function for creating the array of anchor urls, into it new function that is responsible for converting heading text to anchors. This function returns an object containing the anchor id and the url. Since the returns an object I've again renamed the anchorURLs variable to anchorInfo. Also moved the sanitization into an invoker function. This should make it testable and configurable.