FLUID-4209: Step 4 of ToC refactoring from review
Split apart the toModel into two functions. 1) creates an array of heading objects that contains the necessary information, including level, text, and url. 2) transforms the array of heading objects into the hierarchical model. This is still done recursively, but uses an internal function so that it won't destroy the incoming array.