FLUID-4023: Added the Micro Clearfix to FSS
We were unable to obtain an adequate license for the New Clearfix at this moment. Due to this we have switched to the Micro Clearfix. It both has a smaller foot print, and is also licensed under the MIT licenses. I've updated the demo css styles in the fss-layout.css.