FLUID-4209: Continuing unit tests
Started adding tests for the levels subcomponent. There are tests in place for the tree generation. I started writing tests for rendering, but am having trouble with the demands block when trying to set the resource path. Also, even when the resource path is specified in the options, the afterRender event isn't firing. A bunch to look into here.