FLUID-4209: Fixed up the levels tests
Uncommented and fixed up the levels unit tests. I had to set the resource url in the defaults since the levels subcomponent was being called directly. Also had to fix the test case itself a bit. One of the tests is still failing. This is a known issue that I'm still trying to track down. Hopefully this test will make it easier to spot.