FLUID-3761: In progress live preview work
Mike and my in progress work to get live preview functioning. We've got the general structure set up. The integrator will interact with what looks like a top level uiOptions component, from the options perspective. Underneath this is split apart with a portion being rendered into the iframe and part into the parent. This currently isn't working yet, but we wanted to save our progress.