FLUID-3761: A step closer to live preview working
This is as far as triggering the initialization of uiOptions in the iframe. There is a problem with the settings store, likely around the expansion of the uiEnhancer in the tree too early.

I also did some minor refactoring to correct some mispelt names and add in the event binding for saving when the sliding panel hides (this is not factored correctly yet, needs more work still).