FLUID-3761: Adding default background for iframe
If the iframe's body doesn't have a default background colour set, the parent page's background colour will show throw. I had to set it on the body directly, instead of using a class name. I think this is likely due to the order of the class names when using uio to change things.