FLUID-4207: Made updating from the store public.
Added a public function which will allow an integrator to trigger the uiEnhancer to update directly from it's store. This is used internally in the finalInit to load the initial settings (unit tests for this still pass). This will also be used by the fat panel UI Options to update the UIO when the panel is opened.