FLUID-3761: mostly working preview
The UI Options panel will still flicker on the initial opening of the panel when there are changes to the textSize or lineSpacing. This is likely due to the fact that it needs to wait till the panel is opened before it can determine the initial sizes. All subsequent changes and hide/show of the panel work properly.