FLUID-4299: opts for the cookie in the cookieStore
The cookieStore can now take in options such as expires and path on top of the name for the cookie. There is also now a default path specified, which will prevent a bug that had the cookie not accessible when to a parent page when it was set in an iframe (e.g. for fat panel uio).