FLUID-4209: Fixes unit test failure in IE7
Using .attr('href') on <a> in IE7 would return a full URL (not relative) if the element was not originally in the DOM tree.  In otherwords, when the renderer calls "$(node).html(rendered);" in framework/renderer/js/fluidRenderer.js line 1492, the "href" attribute was injected into the DOM tree and IE7 would return a full URL.  This is a known issue and to by pass it, we will stripe the URI location path if it exists in the unit test.