FLUID-4331: removed duplicate class name
Renamed the class added to the container (iframe) so that we wouldn't be using the same class name in two spots. This should help start to unravel the tangles. Also moved the border styling to the new class name. Had to specify the proper container for the tabs, as the styling was somehow dependent on the class added to the iframe element on the parent page.