FLUID-4209: toModel() fixed, all unit tests pass.
toModel() fixed, now properly returning the model depends on the two strategy, one is gradual indentation [h1, h6], the other is skipped indentation, [h1, '', '', '', '', h6].  Unit tests are modified because the generateTree() and toModel() functions are modified.  fluid.tableOfContents.modelBuilder.gradualModelLevelFn() also changed to decrement subHeadings.level by 1.  Note that we have to clone the subHeadings before decrementing it, otherwise it will affect the recursion because of the level reference.