FLUID-4343: Refactored to not have options that have no values.
I've refactored the cookie string assembly. To prevent the issue in IE where the cookie wasn't being set if the expires="" case, the path and expires are only added when there is a value specified for them. I've added new unit tests for the assembly code and fixed the name of the test page.