FLUID-4219: searching a directory for !important injection
Refactored the code to search through a provided directory and attempt to inject !important into all of the files within. There needs to be code to make sure that only css files are used. Also had to hardcode the exclusion of "fss-JSR168Bridge.css" and "fss-transitions.css", which break the !important injection. The first file is empty, so that's probably the issue there, but not sure about the second one.