FLUID-4219: Refactoring to use eval, not done yet
Part way through refactoring to use eval instead of trying to concatenate the js files in ant. This is to make the code more reusable. However, there are issues with using eval in rhino that still need to be worked through. It seems as though we are stuck in some sort of strict mode.