FLUID-4219: added a module for files that need injection
Added a new JSON file to define which css files need the !important inje
ction performed on it. I have currently created a new a new cssModule ty
pe to define this module location in the build.properties file. The curr
ent cssBasePath property should be removed once the code is written to r
ead in this module.