FLUID-4367: Fixing border colours in contrast themes
There were a few things needed to get the border colours to work with
the themes. 1) a couple of the themes were missing the border-color
property for inputs and divs 2) the borders on iframes weren't
considered in the contrast themes, 3) the border-color properties needed
to have !important.  I also added in the missing properties that are
supposed to get !important injection in the generation code.