FLUID-4352: Added unit test for empty headings
Added unit tests as suggested by https://github.com/fluid-project/infusion/pull/122#issuecomment-1683235.  Used some of the existing test functions with different parameter to test empty headings.  Also added a new markup with no headings to test the rendering.  Added a check specifically in renderTOCTest() to ensure <ul> is not in the markup when headings are empty.