FLUID-4366: fixing the styling of the sliders with the contrast themes
The sliders weren't showing up properly when themes were applied because
the jQuery UI themes were not scoped to the new generated themes class
name. To work around this we are adding both classnames when the theme
is chosen. Fat panel also had an issue caused by the overrides to the
jQuery UI tabs. I've scoped these to the .fl-tabs class now.