FLUID-4377: Corrected and renamed the jQuery UI contrast themes
Corrected and renamed the jQuery UI specific contrast themes. The hci
and hc theme names are still in place, but redirect to the wb and bw
themes respectively. This is just for backwards compatibility. Still
need to update the names of the FSS themes and all the references of
them in the code base.