FLUID-4370: Setting contrast styles on all elements
Using the universal selector "*" to set the colour, background colour,
and border colour or all elements. There is still an issue with this for
the "Colour & Contrast" select box which should have theme specific
styling for each option. However these have the same specificity as the
active theme so may be overridden. Need to investigate this more.