FLUID-4370: Moved the select menu overrides to UIO css
Moved the select menu style overrides out of the themes and into UIO's
own css file. I had to add in the "!important" since this file doesn't
get passed through the css generation. All of this is still a hack and
should be removed when a better option is found.