FLUID-4370: Styling Default "Colour & Contrast" theme option
Made some modifications so that the Default option for the "Colour &
Contrast" select box will be styled even when other themes are applied.
To get this to work I had to modify the classnameMap to have a default
theme class and change the clearClasses function used by the
classSwapper to not globally remove theme classes, but only remove it
from its container. Additionally I was able to remove some knowledge of
the themes from the uiOptions.css