FLUID-4370: Moved addition of fl-preview-theme class to UIO and start of fixing IE6 bug
Removed the "fl-preview-theme" from being added by the uiEnhancer and
moved it to UIO specifically for the select box options. Noticed an IE6
bug where the elements with fl-inverted-color are getting the by theme
contrast styling when the default theme is selected. Started to fix this
by creating a small component to just remove the fl-inverted-color
classes in IE6, but it isn't being initialized. There may also be a bug
with ToC in IE6; which seemed to through an error when it was started.