FLUID-4602: Fixed the browseButton selector
The default for the browseButton selector in "fluid.uploader.html5Strategy.local" was set incorrectly. The setting was missing "options" from the EL path. This prevented a user from being able to change the selector for the browseButton from fluid.uploader. I've made the necessary change and verified that the unit tests are still passing.