FLUID-4794: updating progressive enhancement
Adding in two new functions to the progressive enhancement: check and forget. The check function takes an object where the keys will map to the static environment if the value is a function that evaluates to true. This can be used, for example, to test browser features and only set the static environment if they exist. The forget function takes a list of keys to remove from the static environment.