FLUID-4794: first pass at updating uploader PE.
This is the first pass at using the new Progressive Enhancement scheme with uploader. THe progressiveChecker is removed in favour of using demands blocks straight. There will need to be some clean up of the browser check function in Progressive Enhancement as they won't be available at all in non browser environments and will likely through errors instead of just not setting the static environment. The check function has also been updated to not run again if the specified key is already in the static environment.