FLUID-4794: Updated the check function.
Updated the check function to use the fluid.progressiveEnhancement.typeToKey function to ensure that the keys in the static environment are valid. Additionally, it will record all of the keys, and the results of the check into the fluid.progressiveEnhancement.checked object. If a key is already found in the checked object, the functions will not be rerun. The forget function will be updated shortly to make use of the checked object for forgetting single and all checks.