FLUID-4927: Moved default panels config to a grade
Moved the default settings panels configuration into a grade "fluid.uiOptions.defaultSettingsPanels". The intention was to provide this as a gradeName to an instance of UIO that wished to pull in the default settings panels at time of instantiation. However this currently will not work in the framework due to FLUID-4937. In the meantime, since this component itself has fluid.uiOptions as a grade, it is used inplace of fluid.uiOptions when we need the default settings panels. Another possible work around would be to use the dynamically assigned gradeNames that are being worked on in the FLUID-4921 branch.