FLUID-4907: refactored AuxiliaryParser
Refactored the parser code to simplify the expansion functions. This also enabled for the removal of the fluid.uiOptions.auxiliaryExpander function, as the fluid.uiOptions.expandSchema function can be used directly now. Also fixed the bugs in expandSchema that were causing it to generate the incorrect output. Unit test are all passing, but still need an additional one for the fluid.uiOptions.auxBuilder grade.