FLUID-4907: refactored builder and updated tests.
Refactored the builder to be inline with the options returned from the auxiliary builder. For example the options returned will always be top level and will include gradeNames. Also updated the unit tests. Still need to write a few more tests for only panels, everything, and to factor in the messages option from the auxiliary schema.