FLUID-4907: creating consolidated grade
Refactored builder to create a consolidated grade from the created grades, this will enable an integrator to just invoke a single component to create the entire uio/uie stack. All existing tests are passing and added new ones for the new functions. Still need to test the new members for creating the consolidated grade.