FLUID-4907: Updated builder to take templatePrefix
Updated the builder to use the templatePrefix grade. Unit tests updated for this. Also made a small change to the fluid.uiOptions.builder.uio grade to remove the components block. This is because the fluid.uiOptions.inline grade preforms transformations to move these values into the correct spot, so the nesting isn't required.