FLUID-5131: refactored produce tree
Modified the composite panel to override the produceTree invoker instead of the function in the options. This allows an integrator to use the produceTree function in the options still, or a protoTree. The produceTree will need to be fully expanded though. Which is set in the defaults for the compositePanel. Also did some minor refactoring to simplify the expandProtoTree function.