FLUID-5180: Making string lookup easier.
Simplified the process of looking up strings from a message resolver by providing an IoC mechanism for resolving the string from {that}.stringBundles.stringKey. There is also a way to return an array of strings by providing a stringArrayIndex option to the component, which contains an array of stringKeys that should be resolved. One major change that was required for this is to change the prefix separator used for grouping strings in a messageBundle.