FLUID-5177, FLUID-5020, FLUID-4569, FLUID-4410:
Refactored the slidingPanel to make it more declarative and inline with current code style.

FLUID-5177: DOM events are not directly passed down, but the DOM event is bound declaratively. To modify the DOM event object an integrator could either declaratively bind an additional click event, or override the current click handler function.

FLUID-5020, FLUID-4410: By adding a separate selector for the toggleButtonLabel, an integrator can use more complex markup inside of the toggle button. By default the button and the label point at the same element.

FLUID-4569: The slidingPanel is model bound, and has been for some time. Steps have been taken to ensure that the initial state is set to the value specified by the model.