FLUID-5120: Fixed a bug in BuilderTests
The builder tests had been passing by chance because the file containing the implementation for the default prefsEditor type wasn't included. When this was included, the test would break as it wasn't setup to run the SeparatedPanelPrefsEditor. This was resolved by explicitly using the Full No Preview editor, and correcting the template paths.