FLUID-5284: modified jshintrc
changed "forin": true to false.
added a globals config for the fluid version namespace (i.e. fluid_1_5)

Also removed the jshint comment from Gruntfile.js as it is expected that jshint will run with the config from jshintrc.
Note that for non-standard config the jshint and globals comment should be added to the file.