FLUID-5284: fixed another strict mode issue
In some of the tests, fluid.allocateSimpleId would fail. This was the result of element being a string value and an attempt from the function to assign a value onto it's id property. This was corrected by adding an additional check for fluid.isPrimitive.

Also removed an unused css include from one of the manual test files.