FLUID-4698: setting font-size on html
Modified the textSize enactor to set the font-size on the root element (html) instead of the container (body). This will allow support of rem units. Another change with this was to set the value in px instead of ems. This is necessary since the root element has no parent for which to base it's value on. The unit tests have been updated.