FLLUID-4698: Textsize set on correct html now
Fixed an issue where the text size would always be set on the parent document. This was causing issues with the separated panel which uses an iframe for the controls and has its own enactors. Also made another small modification that ensures the multiplier used is always a legitimate value.