FLUID-5339, FLUID-5343: Fixed bug in Databinding
Databinding.js had been using a non-existent function to perform a sort operation on an array. There was a similar function in the file that wasn't being used. Assuming this to be either a typo or refactoring error, I switched the non-existent function for the unused one. The tests are now passing in IE8 and are still working elsewhere.