FLUID-5675: Added MockTTS for integration tests
Added a mock version of fluid.textToSpeech which mocks the calls to the Web Speech API. This will allow for easier integration testing and remove the need for these tests to be run in a browser that supports the Web Speech API (e.g. Firefox).

The unit tests for the selfVoicingEnactor have been updated to make use of the MockTTS.