FLUID-5730: removed qunit-fixture element
The test markup had been inside a qunit-fixture; however, qunit was clearing out
the markup causing the test component to initialize without all of the markup it
was expecting. This was only an issue in IE. I've removed the fixture as there
is only one test that uses the markup.