FLUID-4770: Temporary fix for misalignment.
Added the code that changes the width of the button to the uploader.css in
the uploader demo. Somehow the text size increases by adding
foundation.css and normalise.css so changing the width in css of the
uploader component itself is meaningless as a change in text size would
make the button misaligned again. Good for a temporary fix so that the demo
work fine. Tested in firefox/IE/Edge/Chrome.