FLUID-5927: Asserting that context are the same DOM element
Doing a deep equals on the documents causes Safari to fail as there are some lower level properties
that aren't set yet and have TyperError as a value. E.g. selectionStart. However it should be
sufficient to do an assertEquals to ensure that the contexts refer to the same Document.