FLUID-5980: Adjusting the position of the tooltip
The tooltip was overlapping the input field when keyboard focus was used. This is because it was being
positioned relative to the textEditButton which was set with the hiddenAccessible class. The "of" property
was used to specify that it should be positioned relative to the viewEl in all cases.

Also updated the offset as it is no longer specified in a sperate property. The default value for this has also
been removed from the Tooltip component.