FLUID-6382: Preventing loading the same pattern file multiple times.
In the case where more than one language code points to the same pattern 
file, there could be multiple attempts to load the same pattern if it 
hadn't already been loaded before a new request was submitted. To 
prevent this we keep track of which sources are requested and share the 
promise related to them.