FLUID-6278: Further in progress work.
The domReader is mostly working again. The current issue is that if the 
selection reader is running and we start the domReader, the onStop event 
from the selection reader's utterance triggers the handler in the 
domReader which confuses the state. This results in the highlighting not 
appearing and the button controller state to be incorrect. Also need to 
fix the tests and comments.