FLUID-6420: In progress commit
Auxiliary builder working with the panels/enactors being keyed off of 
preference. Removed the @ expansion from the auxiliary schema. Need to 
devise a better way to pass down preferences to the prefsEditorloader 
and prefsEditor. Currently some prefsEditorLoader options are grouped 
under "prefsEditor" in the auxiliary schema but the name isn't quite 
right and only those options are passed along. Need a way to pass along 
any potential configuration through the schema, or require that those 
options are passed down some other way, and these are more clearly 
defined as relating only to the build config.