FLUID-6431: Updating selection button rendering on text change.
Right Clicks in some browsers also perform a selection. In Safari and 
Chrome those selectionchange events fire twice but with the same text 
selection so there is no time where the UI would have been removed. By 
rendering the text on all selection changes, the UI is always at the 
beginning of the text and will follow right click selections.