NOJIRA: Turning off coverage checks for patches.
The patch check is meant to ensure that changes have a specified level of coverage. However, we found that the overal project coverage was going up by the patch still failing. For example in FLUID-6548 PR that chnaged linting, touched much of the repo, but didn't specifically constitute changes that required further unit tests.